iframe border does not show anymore.
Originally, the border of the iframe was visible. Recently, I learned how to hide that border, so I did it to make the page look clean even though I know that my HTML is very basic and that was the cleanest it was going to get. Yes, I know it's 9:33 p.m. when I'm posting this update.